web: Review changes
Refactor/renam BackendAlertOverlay <-> AlertOverlay
Move AlertOverlay to parent components directory
Css fix: focus-within
Some basic code clean up
