+++ client/web/src/insights/components/alert-overlay/AlertOverlay.module.scss
